feature(Window): Window now gathers inputs that can be used.
The window is in charge of gathering input, which makes sense since it could depend on another windowing library... (which makes my implem invalid, we'd need an extra layer of abstraction with a GLFW implementation, but anyway). The output of these functions will be used by App.
Right now, the only input we handle are directional arrows and mouse. We'll probably need a huge enum to abstract the window lib entirely and only rely on our own types in classes that use Window.