feature(Texture): Fixed Allocate + utility functions
Allocate had 2 variables reversed. Added more options to store the internal format after texture construction, to send an Image to a shader, to set border color on wrap border, and a new internalFormat function that is horible right now, but there is SO MANY formats i'm being lazy... It should be its own enum.